pyproject.toml
[tool.poetry]
name = "protomesh-py"
version = "0.1.0"
description = "Protomesh Python SDK"
authors = ["Victor França Lopes <18534956+vflopes@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "protomesh_py"}]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.56.2"
aws-lambda-typing = "^2.17.1"
protobuf = "^4.23.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
